BrowseFragment now requests a data file
if the 'library' hasn't been loaded from the server, fire up a
JSONRequestTask and fetch it. Once fetched, process it on the UI
thread, updating the database accordingly. In tests, this is
a very short time. With the full size stop/tag library, this could
take quite a while.

Basically, it may have to be moved to a separate thread, then we
only have to worry about concurrent database access. For now, I'll
keep it simpler.

Also, went ahead and made WtaFragment implement Start/Stop progress
because all views now may need to show progress and errors.