NextBusFragment is now responsible for stop info
Adds a lot of unfortunate complexity to the WtaFragment however. There
is some refactoring necessary, but for now this is a step in the right
direction (i think).

Also safeguarded against the triggering of an infrequent but common bug
which crashes the app on start if it hasn't been used in a while. I'll
be revisiting that one as well, but for now the app is more stable (i
think).